refactor: move search_indicators_enriched to api.py
- Moved _resolve_country_code and search_indicators_enriched from tools.py to api.py
- Simplified tools.py to thin wrapper layer (~50 lines)
- All business logic now in api.py, tools.py just registers MCP tools
